Set primary wallet Signed-off-by: Marc Juchli <marc.juchli@digitalasset.com>
