+++ core/wallet-user-rpc-client/package.json
@@ -21,14 +21,14 @@
     },
     "devDependencies": {
         "@eslint/js": "9.21.0",
+        "@typescript-eslint/eslint-plugin": "^8.25.0",
+        "eslint-config-prettier": "^10.0.1",
+        "eslint-plugin-prettier": "^5.2.3",
         "@types/isomorphic-fetch": "^0.0.39",
         "@types/jest": "^29.5.12",
         "@types/json-schema": "7.0.3",
         "@types/lodash": "^4.14.149",
         "@types/ws": "^6.0.1",
-        "@typescript-eslint/eslint-plugin": "^8.25.0",
-        "eslint-config-prettier": "^10.0.1",
-        "eslint-plugin-prettier": "^5.2.3",
         "globals": "^16.0.0",
         "prettier": "^3.5.2",
         "typedoc": "^0.27.9",
