+++ core/wallet-user-rpc-client/src/index.ts
@@ -193,6 +193,10 @@ export interface CreateWalletParams {
     signingProviderId: SigningProviderId
     [k: string]: any
 }
+export interface SetPrimaryWalletParams {
+    partyId: PartyId
+    [k: string]: any
+}
 export interface RemoveWalletParams {
     partyId: PartyId
     [k: string]: any
@@ -276,6 +280,7 @@ export type RemoveNetwork = (params: RemoveNetworkParams) => Promise<Null>
 export type CreateWallet = (
     params: CreateWalletParams
 ) => Promise<CreateWalletResult>
+export type SetPrimaryWallet = (params: SetPrimaryWalletParams) => Promise<Null>
 export type RemoveWallet = (
     params: RemoveWalletParams
 ) => Promise<RemovePartyResult>
@@ -321,6 +326,15 @@ export class SpliceWalletJSONRPCUserAPI {
         method: 'createWallet',
         ...params: Parameters<CreateWallet>
     ): ReturnType<CreateWallet>
+
+    /**
+     *
+     */
+    // tslint:disable-next-line:max-line-length
+    public async request(
+        method: 'setPrimaryWallet',
+        ...params: Parameters<SetPrimaryWallet>
+    ): ReturnType<SetPrimaryWallet>
 
     /**
      *
